sources/kafka_dlt/requirements.txt
kafka-python>=2.0.0
dlt>=0.3.5,<0.4
